# openapi.yaml
openapi: 3.1.0
info:
  title: Waternity API
  description: |
    Waternity is a transparent water sales and settlement platform built on Hedera Hashgraph.
    This API provides endpoints for managing water projects, wells, settlements, and audit trails.
    
    ## Authentication
    The API uses cookie-based authentication with httpOnly JWT tokens. After successful login,
    a secure session cookie (`wty_sess`) is set with 8-hour expiry.
    
    ## Role-Based Access Control
    - **INVESTOR**: Can invest in projects and view settlements
    - **OPERATOR**: Can create wells/projects, manage operations, control meters/valves
    - **AGENT**: Can verify/approve settlements and anchor documents
    - **ADMIN**: Full system access
    
    ## Hedera Integration
    All operations are recorded on Hedera Consensus Service (HCS) for immutable audit trails.
    Links to HashScan explorer and Mirror Node are provided for verification.
  version: 1.0.0
  contact:
    name: Waternity Team
    url: https://waternity.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://waternity.vercel.app
    description: Production server
paths:
...
  /api/wells:
    get:
      tags:
        - Wells
      summary: List all wells
      description: |
        Retrieve a list of all water wells in the system.
        This endpoint is publicly accessible for transparency.
      operationId: listWells
      parameters:
        - name: limit
          in: query
          description: Maximum number of wells to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Number of wells to skip
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: status
          in: query
          description: Filter by well status
          schema:
            type: string
            enum: [active, inactive, maintenance]
      responses:
        '200':
          description: Wells retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WellsListResponse'
              examples:
                success:
                  value:
                    wells:
                      - id: "well_001"
                        name: "Sumur Desa Makmur"
                        location:
                          address: "Desa Makmur, Kecamatan Sejahtera"
                          latitude: -6.2088
                          longitude: 106.8456
                        status: "active"
                        tokenId: "0.0.123456"
                        topicId: "0.0.789012"
                        links:
                          hashscan: "https://hashscan.io/testnet/topic/0.0.789012"
                          mirror: "https://testnet.mirrornode.hedera.com/api/v1/topics/0.0.789012/messages"
                    count: 1
                    total: 1
        '500':
          $ref: '#/components/responses/InternalError'
  /api/wells/{id}/events:
    get:
      tags:
        - Wells
      summary: Get well events
      description: |
        Retrieve all events for a specific well from the audit trail.
        Events are sourced from Hedera Consensus Service.
      operationId: getWellEvents
      parameters:
        - name: id
          in: path
          required: true
          description: Well ID
          schema:
            type: string
            example: "well_001"
        - name: limit
          in: query
          description: Maximum number of events to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: eventType
          in: query
          description: Filter by event type
          schema:
            type: string
            enum: [project_created, well_completed, meter_reading, valve_command, settlement_requested]
      responses:
        '200':
          description: Well events retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WellEventsResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'

  # Settlements
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: wty_sess
      description: |
        Session-based authentication using httpOnly JWT cookies.
        The cookie is automatically set after successful login and cleared on logout.
        
        Cookie attributes:
        - HttpOnly: true (prevents XSS attacks)
        - Secure: true (HTTPS only in production)
        - SameSite: Lax (CSRF protection)
        - Max-Age: 28800 (8 hours)
  schemas:
...
    WellsListResponse:
      type: object
      properties:
        wells:
          type: array
          items:
            $ref: '#/components/schemas/Well'
        count:
          type: integer
          description: Number of wells in current response
          example: 20
        total:
          type: integer
          description: Total number of wells
          example: 150
    Well:
      type: object
      properties:
        id:
          type: string
          example: "well_001"
        name:
          type: string
          example: "Sumur Desa Makmur"
        location:
          $ref: '#/components/schemas/Location'
        status:
          type: string
          enum: [active, inactive, maintenance]
          example: "active"
        tokenId:
          type: string
          description: Hedera token ID
          example: "0.0.123456"
        topicId:
          type: string
          description: Hedera topic ID for events
          example: "0.0.789012"
        links:
          $ref: '#/components/schemas/HederaLinks'
    WellEventsResponse:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/WellEvent'
        count:
          type: integer
          example: 25
    WellEvent:
      type: object
      properties:
        messageId:
          type: string
          format: uuid
        eventType:
          type: string
          example: "meter_reading"
        timestamp:
          type: string
          format: date-time
        triggeredBy:
          type: string
          pattern: "^0\\.0\\.[0-9]+$"
        eventData:
          type: object
          description: Event-specific data
        hederaData:
          $ref: '#/components/schemas/HederaTransaction'

    # Settlement Schemas
...
    Location:
      type: object
      required:
        - latitude
        - longitude
      properties:
        address:
          type: string
          maxLength: 200
          example: "Desa Makmur, Kecamatan Sejahtera, Kabupaten Maju"
        latitude:
          type: number
          minimum: -90
          maximum: 90
          example: -6.2088
        longitude:
          type: number
          minimum: -180
          maximum: 180
          example: 106.8456
...
    HederaTransaction:
      type: object
      properties:
        transactionId:
          type: string
          description: Hedera transaction ID
          example: "0.0.12345@1705315800.123456789"
        consensusTimestamp:
          type: string
          description: Consensus timestamp from Hedera
          example: "1705315800.123456789"
        hashscanUrl:
          type: string
          format: uri
          description: URL to HashScan explorer
          example: "https://hashscan.io/testnet/transaction/0.0.12345@1705315800.123456789"
    HederaLinks:
      type: object
      properties:
        hashscan:
          type: string
          format: uri
          description: HashScan explorer link
          example: "https://hashscan.io/testnet/topic/0.0.789012"
        mirror:
          type: string
          format: uri
          description: Mirror Node API link
          example: "https://testnet.mirrornode.hedera.com/api/v1/topics/0.0.789012/messages"
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
        timestamp:
          type: string
          format: date-time
  responses:
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            notFound:
              value:
                error: "not_found"
                message: "Resource not found"
                timestamp: "2024-01-15T10:30:00Z"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            internal:
              value:
                error: "internal_error"
                message: "An internal error occurred"
                timestamp: "2024-01-15T10:30:00Z"
...